+++ src/hearings/models/hearingResponse.model.ts
@@ -1,13 +1,1 @@
-import { HearingDayScheduleModel } from './hearingDaySchedule.model';
-import { HearingListingStatusEnum, LaCaseStatus } from './hearings.enum';
-
-export interface HearingResponseModel {
-  listAssistTransactionID: string;
-  receivedDateTime: string;
-  responseVersion: number;
-  laCaseStatus: LaCaseStatus;
-  listingStatus: HearingListingStatusEnum;
-  hearingCancellationReason: string;
-  hearingDaySchedule: HearingDayScheduleModel[];
-  errorTimestamp?: string;
-}
+export { HearingResponseModel } from './hearing-payload.model';
