+++ src/hearings/models/individualDetails.model.ts
@@ -1,17 +1,1 @@
-import { RelatedPartiesModel } from './relatedParties.model';
-
-export interface IndividualDetailsModel {
-  title?: string;
-  firstName?: string;
-  lastName?: string;
-  preferredHearingChannel?: string;
-  interpreterLanguage?: string;
-  reasonableAdjustments?: string[];
-  vulnerableFlag?: boolean;
-  vulnerabilityDetails?: string;
-  hearingChannelEmail?: string[];
-  hearingChannelPhone?: string[];
-  relatedParties?: RelatedPartiesModel[];
-  custodyStatus?: string;
-  otherReasonableAdjustmentDetails?: string;
-}
+export { IndividualDetailsModel } from './hearing-payload-support.model';
