+++ src/hearings/models/judiciary.model.ts
@@ -1,15 +1,1 @@
-import { PanelPreferenceModel } from './panelPreference.model';
-
-export interface JudiciaryModel {
-  roleType: string[];
-  authorisationTypes: string[];
-  authorisationSubType: string[];
-  panelComposition: [
-    {
-      memberType: string;
-      count: number;
-    },
-  ];
-  judiciaryPreferences: PanelPreferenceModel[];
-  judiciarySpecialisms: string[];
-}
+export { JudiciaryModel } from './hearing-payload-support.model';
