+++ src/hearings/models/panelRequirements.model.ts
@@ -1,9 +1,1 @@
-import { PanelPreferenceModel } from './panelPreference.model';
-
-export interface PanelRequirementsModel {
-  roleType?: string[];
-  authorisationTypes?: string[];
-  authorisationSubType?: string[];
-  panelPreferences?: PanelPreferenceModel[];
-  panelSpecialisms?: string[];
-}
+export { PanelRequirementsModel } from './hearing-payload-support.model';
