+++ src/hearings/models/partyUnavilabilityRange.model.ts
@@ -1,4 +1,4 @@
-export interface PartyUnavailabilityRange {
-  start: string;
-  end: string;
-}
+export { PartyUnavailabilityRangeModel } from './hearing-payload-support.model';
+
+// Preserve the legacy symbol exported by this misspelled path.
+export { PartyUnavailabilityRangeModel as PartyUnavailabilityRange } from './hearing-payload-support.model';
