+++ src/hearings/models/requestDetails.model.ts
@@ -1,12 +1,1 @@
-export interface ResponseDetailsModel {
-  hearingRequestID?: string;
-  status?: string;
-  timestamp: string;
-  versionNumber: number;
-}
-
-export interface RequestDetailsModel extends ResponseDetailsModel {
-  hearingGroupRequestId?: string;
-  partiesNotified?: string;
-  cancellationReasonCodes?: string[];
-}
+export { RequestDetailsModel, ResponseDetailsModel } from './hearing-payload.model';
