+++ src/hearings/models/screenNavigation.model.ts
@@ -1,11 +1,1 @@
-export interface NavigationModel {
-  conditionOperator?: string;
-  conditionValue?: string;
-  resultValue: string;
-}
-
-export interface ScreenNavigationModel {
-  screenName: string;
-  conditionKey?: string;
-  navigation: NavigationModel[];
-}
+export { NavigationModel, ScreenNavigationModel } from './hearing-payload-support.model';
