+++ src/hearings/models/unavailabilityRange.model.ts
@@ -1,7 +1,1 @@
-import { UnavailabilityType } from './hearings.enum';
-
-export interface UnavailabilityRangeModel {
-  unavailableFromDate: string;
-  unavailableToDate: string;
-  unavailabilityType: UnavailabilityType;
-}
+export { UnavailabilityRangeModel } from './hearing-payload-support.model';
